+++ Command/ImportMapRequireCommand.php
@@ -36,12 +36,8 @@ final class ImportMapRequireCommand extends Command
     public function __construct(
         private readonly ImportMapManager $importMapManager,
         private readonly ImportMapVersionChecker $importMapVersionChecker,
-        private readonly ?string $projectDir = null,
+        private readonly string $projectDir,
     ) {
-        if (null === $projectDir) {
-            trigger_deprecation('symfony/asset-mapper', '7.3', 'The "%s()" method will have a new `string $projectDir` argument in version 8.0, not defining it is deprecated.', __METHOD__);
-        }
-
         parent::__construct();
     }
 
@@ -149,9 +145,7 @@ EOT
                 array_map(fn (ImportMapEntry $package): array => [
                     $package->importName,
                     $package->version ?? '-',
-                    // BC layer for AssetMapper < 7.3
-                    // When `projectDir` is not null, we use the absolute path of the package
-                    null !== $this->projectDir ? Path::makeRelative($package->path, $this->projectDir) : $package->path,
+                    Path::makeRelative($package->path, $this->projectDir),
                 ], $newPackages),
             );
         }